Add some new features

+++ app/Http/Controllers/CharacterController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Character;
 use App\Account;
 use App\Group;
+use App\Town;
 
 class CharacterController extends Controller
 {
@@ -28,9 +29,15 @@ class CharacterController extends Controller
      */
     public function create()
     {
-        $groups = Group::pluck('name', 'id');
-        $accounts = Account::pluck('name', 'id');
-        return view('admin.characters.create', compact('accounts', 'groups'));
+        $towns = Town::pluck('name', 'id');
+
+        if ($this->isAdminRequest()) {
+            $accounts = Account::pluck('name', 'id'); 
+            $groups = Group::pluck('name', 'id');
+            return view('admin.characters.create', compact('accounts', 'groups', 'towns'));                  
+        }
+
+        return view('site.characters.create', compact('towns'));
     }
 
     /**
@@ -42,10 +49,23 @@ class CharacterController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, Character::rules());
-        
-        Character::create($request->all());
 
-        return redirect('admin/characters')->withSuccess(trans('admin.success_store'));
+        if ($this->isAdminRequest()) {
+            if (!Character::create($request->all())) {
+                return view($view)->withErrors(trans('admin.success_store'));
+            }
+
+            return view($view)->withSuccess(trans('admin.success_store'));
+        }
+
+        $character = new Character($request->all());
+        $character->account_id = auth()->user()->id;
+
+        if (!$character->save()) {
+            return redirect()->route('accounts.show')->withErrors(trans('admin.success_store'));
+        }
+
+        return redirect()->route('accounts.show')->withSuccess(trans('admin.success_store'));
     }
 
     /**
@@ -102,8 +122,40 @@ class CharacterController extends Controller
      */
     public function destroy($id)
     {
-        Character::destroy($id);
+        if ($this->isAdminRequest()) {
+            Character::destroy($id);
+            return back()->withSuccess(trans('admin.success_destroy'));
+        }
 
-        return back()->withSuccess(trans('admin.success_destroy')); 
+        $character = Character::find($id);
+
+        if ($character->account_id != auth()->user()->id) {
+            return back()->withErrors(trans('admin.error_destroy'));
+        }
+
+        $character->deletion = now()->addDays(10)->timestamp;
+        $character->save();
+
+        return back()->withSuccess(trans('admin.success_destroy'));
+    }
+
+    public function undestroy($id)
+    {
+        $character = Character::find($id);
+
+        if ($this->isAdminRequest()) {
+            $character->deletion = 0;
+            $character->save();
+            return back()->withSuccess(trans('admin.success_destroy'));
+        }
+
+        if ($character->account_id != auth()->user()->id) {
+            return back()->withErrors(trans('admin.error_destroy'));
+        }
+
+        $character->deletion = 0;
+        $character->save();
+
+        return back()->withSuccess(trans('admin.success_destroy'));   
     }
 }
